+++ src/TypeParser/PhpTypeParser.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Liip\MetadataParser\TypeParser;
 
 use Doctrine\Common\Annotations\PhpParser;
+use Doctrine\Common\Collections\Collection;
 use Liip\MetadataParser\Exception\InvalidTypeException;
 use Liip\MetadataParser\Metadata\PropertyType;
 use Liip\MetadataParser\Metadata\PropertyTypeArray;
@@ -56,14 +57,25 @@ final class PhpTypeParser
             }
         }
 
-        if (0 === \count($types)) {
+        $isCollection = false;
+        $filteredTypes = [];
+        foreach ($types as $type) {
+            $resolvedClass = $this->resolveClass($type, $declaringClass);
+            if (is_a($resolvedClass, Collection::class, true)) {
+                $isCollection = true;
+            } else {
+                $filteredTypes[] = $type;
+            }
+        }
+
+        if (0 === \count($filteredTypes)) {
             return new PropertyTypeUnknown($nullable);
         }
-        if (\count($types) > 1) {
+        if (\count($filteredTypes) > 1) {
             throw new InvalidTypeException(sprintf('Multiple types are not supported (%s)', $rawType));
         }
 
-        return $this->createType($types[0], $nullable, $declaringClass);
+        return $this->createType($filteredTypes[0], $nullable, $declaringClass, $isCollection);
     }
 
     /**
@@ -78,7 +90,7 @@ final class PhpTypeParser
         throw new InvalidTypeException(sprintf('No type information found, got %s but expected %s', \ReflectionType::class, \ReflectionNamedType::class));
     }
 
-    private function createType(string $rawType, bool $nullable, \ReflectionClass $reflClass = null): PropertyType
+    private function createType(string $rawType, bool $nullable, \ReflectionClass $reflClass = null, bool $isCollection = false): PropertyType
     {
         if (self::TYPE_ARRAY === $rawType) {
             return new PropertyTypeArray(new PropertyTypeUnknown(false), false, $nullable);
@@ -87,12 +99,12 @@ final class PhpTypeParser
         if (self::TYPE_ARRAY_SUFFIX === substr($rawType, -\strlen(self::TYPE_ARRAY_SUFFIX))) {
             $rawSubType = substr($rawType, 0, \strlen($rawType) - \strlen(self::TYPE_ARRAY_SUFFIX));
 
-            return new PropertyTypeArray($this->createType($rawSubType, false, $reflClass), false, $nullable);
+            return new PropertyTypeArray($this->createType($rawSubType, false, $reflClass), false, $nullable, $isCollection);
         }
         if (self::TYPE_HASHMAP_SUFFIX === substr($rawType, -\strlen(self::TYPE_HASHMAP_SUFFIX))) {
             $rawSubType = substr($rawType, 0, \strlen($rawType) - \strlen(self::TYPE_HASHMAP_SUFFIX));
 
-            return new PropertyTypeArray($this->createType($rawSubType, false, $reflClass), true, $nullable);
+            return new PropertyTypeArray($this->createType($rawSubType, false, $reflClass), true, $nullable, $isCollection);
         }
 
         if (self::TYPE_RESOURCE === $rawType) {