+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 0.6.0 (unreleased)
 
 * When running with PHP 8, process attributes in addition to the phpdoc annotations.
+* Support doctrine collections
+* Support `identical` property naming strategy
 
 # 0.5.0